Done with profile & worked on index
i have made a delete user buton that shows a popup, it is mostly working but need a bit of fixing .i also made an popup for loging out. i have alos opemized the js of profile a bit and comented in it.

I have fihished the leave chatroom function and mad it so an user can not create an chatroom with no name , if done it give an popup that informs the user of the problem